苍穹外卖 根据id查询员工代码 更改id数据类型 编辑员工信息代码

+++ sky-server/src/main/java/com/sky/controller/admin/EmployeeController.java
@@ -134,7 +134,7 @@ public class EmployeeController {
      * @return
      */
     @GetMapping("/{id}")
-    public Result<Employee> getById(@PathVariable Integer id) {
+    public Result<Employee> getById(@PathVariable Long id) {
         log.info("根据id查询员工:{}",id);
 
         Employee employee = employeeService.getById(id);
